+++ tests/test_useful_links.py
@@ -1,5 +1,5 @@
 import pytest
-from pages.link_screens import printUsefulLinkScreen, usefulLinksOptionsList
+from pages.useful_links import printUsefulLinkScreen, usefulLinksOptionsList
 
 # underConstructionMessage = "under construction, input anything to return" #
 from common_utils.messages import underConstructionMessage
